+++ src/app/page.module.css
@@ -1,0 +1,215 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  min-height: 100vh;
+}
+/* nosotros primera parte */
+
+.container_nosotros {
+  max-width: 1200px;
+  width: 80%;
+  margin: auto;
+  display: flex;
+  background-color: #E6B54A;
+}
+
+.img_nosotros {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.parrafo_nosotros {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  font-style: normal;
+  padding: 50px;
+  width: 100%;
+  max-width: 400px;
+  color: #FFFFFF;
+}
+
+.parrafo_nosotros h1 {
+  font-family: 'Bebas Neue';
+  font-style: normal;
+  font-weight: 500;
+  font-size: 43px;
+  text-align: center;
+  margin: 0;
+}
+
+.parrafo_nosotros h2 {
+  font-family: 'Bebas Neue';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 30px;
+  margin: 0;
+  text-align: center;
+}
+
+.parrafo_nosotros p {
+  font-family: 'Abel';
+  font-style: normal;
+  font-weight: 180;
+  font-size: 19px;
+  text-align: center;
+}
+
+.saber_mas {
+  position: relative;
+  font-family: 'Bebas Neue';
+  font-style: normal;
+  font-weight: 180;
+  font-size: 22px;
+  text-align: right;
+}
+
+.saber_mas::after {
+  content: '';
+  position: absolute;
+  width: 0;
+  height: 2px;
+  background-color: white;
+  bottom: -3px;
+  left: 280px;
+  transition: width 0.3s ease;
+}
+
+.saber_mas:hover::after {
+  width: 25%;
+}
+
+
+
+
+/*********Contacto*************/
+
+
+.contacto {
+  height: 100vh;
+  min-height: 100vh;
+  /*Letra para toda la sección*/
+  font-family: "Bebas Neue";
+  width: 1500px;
+  max-width: 100%;
+  margin: auto;
+  color: #E6B54A;
+  /*Para que los patrones y la columna central queden en la misma línea*/
+  display: flex;
+  justify-content: space-between;
+  /*Para que los patrones de los costados se escondan*/
+  overflow: hidden;
+  background-color: #FFFFFF;
+}
+
+
+/*Contacto y Ubicación, datos de la columna central*/
+
+.contacto .columna_contacto {
+  display: flex;
+  flex-direction: column;
+  margin: auto;
+  gap: 5px;
+  height: 98%;
+  width: 70%;
+}
+
+
+/*Parte de arriba de la columna, imagen de araña más títulos*/
+
+#columnaarriba {
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+  height: 30vh;
+  margin: 2% auto 2% auto;
+}
+
+
+/*Caja de imagen de araña*/
+
+.imagen_contacto {
+  text-align: center;
+}
+
+
+/*Caja que contiene los títulos*/
+
+.titulos_contacto {
+  text-align: center;
+  margin: 10% auto;
+}
+
+
+/*Fuente de títulos*/
+
+.titulos_contacto h1 {
+  font-size: 6vh;
+}
+
+
+/*Este componente es de datos cliente y mapa*/
+
+.datoscliente_ubicacion {
+  display: flex;
+  flex-direction: row;
+  width: 140vh;
+  max-width: 100%;
+  margin: 5% auto;
+}
+
+
+/*Datos cliente*/
+
+#datoscliente {
+  width: 100%;
+  text-align: center;
+  align-self: center;
+}
+
+
+/*Fuente h1 dentro de datos cliente*/
+
+.contacto .columna_contacto .datoscliente_ubicacion #datoscliente h1 {
+  font-size: 4vh;
+  line-height: 7vh;
+}
+
+
+/*Fuente h3 dentro de datos cliente*/
+
+.contacto .columna_contacto .datoscliente_ubicacion #datoscliente h3 {
+  font-size: 2.5vh;
+  line-height: 3vh;
+  color: rgba(73, 70, 68, 1);
+}
+
+
+/*Fuente de enlaces tipo "a" dentro de datos cliente*/
+
+.contacto .columna_contacto .datoscliente_ubicacion #datoscliente a {
+  font-size: 2.5vh;
+  color: rgba(73, 70, 68, 1);
+  text-decoration: none;
+}
+
+
+/*Fuente de Hover de enlaces tipo "a" dentro de datos cliente*/
+
+.contacto .columna_contacto .datoscliente_ubicacion #datoscliente a:hover {
+  color: #52A5A0;
+  text-decoration: none;
+}
+
+
+/*Mapa*/
+
+#mapa_contacto {
+  padding: 3%;
+  padding: 7px;
+  width: 500px;
+  justify-self: right;
+}
